Add profile.py; Change log file name

+++ utils.py
@@ -47,7 +47,7 @@ def accuracy(output, labels):
 
 
 def normalize(mx):
-    rowsum = np.array(mx.sum(1))
+    rowsum = np.array(mx.sum(1), dtype=float)
     r_inv = np.power(rowsum, -1).flatten()
     r_inv[np.isinf(r_inv)] = 0.
     r_mat_inv = ssp.diags(r_inv)
